backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-docx
python-pptx
pymupdf
groq
pydantic>=2.0.0
pydantic-settings
python-multipart
python-dotenv
